+++ sdk/src/main/java/com/checkout/sdk/uicommon/TextInputUiState.kt
@@ -1,4 +1,4 @@
-package com.checkout.sdk.cvvinput
+package com.checkout.sdk.uicommon
 
 import com.checkout.sdk.architecture.UiState
