+++ sdk/src/main/java/com/checkout/sdk/uicommon/TextInputUseCase.kt
@@ -1,7 +1,6 @@
-package com.checkout.sdk.cvvinput
+package com.checkout.sdk.uicommon
 
 import com.checkout.sdk.architecture.UseCase
-import com.checkout.sdk.uicommon.TextInputStrategy
 
 
 open class TextInputUseCase (
